refactor: Improve import organization in tag_command
